+++ src/State.php
@@ -25,6 +25,10 @@ final class State
 
     public function __construct(Activity $activity, Actor $actor, string $stateId, string $registrationId = null)
     {
+        if (!$actor instanceof Agent) {
+            @trigger_error(sprintf('Passing an instance of "%s" as the second argument is deprecated since 1.2. In 4.0, only instances of "Xabbuh\XApi\Model\Agent" will be accepted.', get_class($actor)), E_USER_DEPRECATED);
+        }
+
         $this->activity = $activity;
         $this->actor = $actor;
         $this->stateId = $stateId;
@@ -41,8 +45,22 @@ final class State
 
     /**
      * Returns the actor.
+     *
+     * @deprecated since 1.2, to be removed in 4.0
      */
     public function getActor(): Actor
+    {
+        @trigger_error(sprintf('The "%s()" method is deprecated since 1.2 and will be removed in 4.0, use "%s::getAgent()" instead.', __METHOD__, __CLASS__), E_USER_DEPRECATED);
+
+        return $this->getAgent();
+    }
+
+    /**
+     * Returns the agent.
+     *
+     * @return Actor The agent
+     */
+    public function getAgent()
     {
         return $this->actor;
     }